refactor: move get_dataloader to common.py, share between train and evaluate
- Move get_dataloader from train.py to datasets/common.py with explicit
  args parameter (no global state dependency)
- evaluate.py uses shared get_dataloader instead of manual build_datasets + DataLoader
- Fix duplicate manifest in train.py (train + val collected same data)
